Blueprints – Functional build_app and cli.build_app (#805)
Replace class-based AppBuilder/CliBuilder with functional blueprints pattern.

- Delete tiferet/builders/ package
- Add tiferet/blueprints/ with build_app, cli.build_app, and tests
- Rename assets/builders.py to assets/blueprints.py (a.bildr -> a.bps)
- Update tiferet/__init__.py exports: App = build_app, CLI = build_cli
- Update pyproject.toml package include
- Fix create_service_provider to register constants before services
- Update integration tests to use functional API
