+++ tiferet/assets/__init__.py
@@ -9,4 +9,4 @@ from .constants import (
     DEFAULT_ERRORS,
 )
 from . import constants as const
-from . import builders as bildr  # + todo: add this to the exports once it's implemented
+from . import blueprints as bps
