+++ tiferet/assets/blueprints.py
