+++ tiferet/blueprints/__init__.py
@@ -1,0 +1,7 @@
+"""Tiferet Blueprint Exports"""
+
+# *** exports
+
+# ** app
+from .main import build_app, build_app as App
+from .cli import build_app as build_cli, build_app as CLI
