+++ tiferet/blueprints/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tiferet Blueprint Tests"""
