--- tiferet/builders/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Tiferet Builder Tests"""
