+++ tiferet/tests_int/test_cli.py
@@ -330,9 +330,8 @@ def test_calc_cli_script(tmp_path, test_calc_yaml_file):
     script = (
         "from tiferet import CLI\n"
         "\n"
-        "cli = CLI().load_app_service(app_yaml_file={main_yaml!r})\n"
         "if __name__ == '__main__':\n"
-        "    cli.run('test_calc_cli')\n"
+        "    CLI('test_calc_cli', app_yaml_file={main_yaml!r})\n"
     ).format(main_yaml=test_calc_yaml_file)
 
     # Write the CLI script to a temporary file.
